Moved none-ls to lsp-config.lua.
By trial and error, found out I dont need the ensure_installed from
mason-none-ls, but for formatters to be installed, it does require
automatic_installation to be true. And for some reason, the LSP's have to be
listed in mason-lspconfig and the formatters in mason.nvim. With it set-up this
way, it will auto-install lsp's and formatters on boot if not already installed.