More lsp-config tweaks.
Figured out that formatters don't have to be listed under ensure_installed in
setup of mason-null-ls. At this point, Im not sure whats the purpose of
ensure_installed is here. Because if 'automatic_installation' is true, and
ensure_installed is empty, it will install the formatters that are being setup
in require("null-ls").setup.

So I guess for auto-install of LSP's and formatters, LSP's need to be listed in
mason-lspconfig. And for formatters, if 'automatic_installation` is true, it
will install formatters setup in the require("null-ls").setup.