Rework lsp's for nvim 0.11
Updated lsp's to work natively. Mason is still used to download and
update lsp's. Mason and nvim-lspconfig are setup in mason-lspconfig.lua.
Lsp's are enabled and configured in config/lsp.lua.

As of right now I perfer using my current plugins for completions, but
if you want to give native completions a try:
 - delete completions.lua
 - uncomment the lines at the bottom of config/lsp.lua
   This will enable native completions.

Unsure how this affects formatters, but they still seem to work for me.